requirements.txt
PySide6>=6.6
pyserial>=3.5
pyqtgraph>=0.13
numpy>=1.24
pandas>=2.0
